Update p1.c
Changed modulus output to use float for division.

+++ p1.c
@@ -14,7 +14,8 @@ int main()
     printf("Multiplication = %d\n", a * b);
     printf("Division = %d\n", a / b);
 
-    printf("Modulus = %d\n", a % b);
+    printf("Modulus = %f\n", (float)(a % b));
 
     return 0;
+
 }
